% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_numeric_cols.R
\name{select_numeric_cols}
\alias{select_numeric_cols}
\title{Select Numeric Columns}
\usage{
select_numeric_cols(dataset, remove_na = FALSE)
}
\arguments{
\item{dataset}{A vector, matrix, data frame, or tibble.}

\item{remove_na}{Logical. If TRUE, rows with any NA values will be dropped. Default is FALSE.}
}
\value{
A tibble with only numeric columns.
}
\description{
Returns a tibble with only the numeric columns of the input, and optionally drops rows with NAs.
}
\examples{
select_numeric_cols(iris)
}
